+++ docs/release-notes.rst
@@ -1,2 +1,14 @@
 Release notes
 =============
+
+1.0.0
+-----
+
+What's new?
+^^^^^^^^^^^
+* Added Identity & access management support - `#1 <https://github.com/dotnet-essentials/Kwality.UVault/issues/1>`_.
+
+Contributors
+^^^^^^^^^^^^
+
+- `Kevin De Coninck <https://github.com/kdeconinck>`_.